pyproject.toml

[project]
name = "glint"
version = "0.1.0"
description = "Your personal and private tech watch assistant"
authors = [
    {name = "Jonathan KABONGA NYATA", email = "jonathan.nyata@gmail.com"},
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
keywords = ["tech", "watch", "notifications", "cli","local"]
dependencies = [
    "click>=8.0.0",
    "requests>=2.25.0",
    "python-dateutil>=2.8.0",
    "rich>=13.0.0", #interface CLI riches
    "typer>=0.9.0",
    "questionary>=2.0.0",
    "sqlmodel>=0.0.8",
    "customtkinter>=5.0.0",
    "packaging>=23.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=21.0",
    "flake8>=3.9",
    "mypy>=0.900",
    "pystest-mock>=3.0",
]

gui = [
    "customtkinter>=5.0.0",
    "tkinter",
]

ml = [
    "scikit-learn>=1.0.0",
    "nltk>=3.6.0",
]

[project.scripts]
glint = "glint.cli.main:app"

[project.urls]
Homepage = "https://github.com/thiszenon/glint"
Documentation = "https://github.com/thiszenon/glint#readme"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
